% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bond_price.R
\name{bond_price}
\alias{bond_price}
\title{Bond Price Calculation}
\usage{
bond_price(face_value = 1000, int_rate, coupon_rate, period)
}
\arguments{
\item{face_value}{Face Value of the Bond, the default value is 1000,}

\item{int_rate}{Interest Rate,}

\item{coupon_rate}{Coupon Payment Rate of the Bond,}

\item{period}{Period of Investment.}
}
\value{
A list with two components

\code{coupon_payment} the coupon payment subject to the interset rate.

\code{bond_price} the bond price.
}
\description{
This function calcuates the value of a bond as a function of the  face value, interest rate (YTM), coupon rate, and the period of investment.
}
\details{
The formula used for the calculation of the bond price is

\deqn{B_p=c\frac{1-(\frac{1}{(1+i)^n}}{i}+\frac{M}{(1+i)}}
where c is the coupon payment, i is the interest rate, M is the face value, and n is the period of investment.
}
\examples{
bond_price(face_value=1000, int_rate=0.17, coupon_rate=0.16, period=6)
}
\references{
Fabozzi Frank J. and Mann Steven V., Handbook of Fixed Income Securities.
}

