% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/macaulay_duration.R
\name{macaulay_duration}
\alias{macaulay_duration}
\title{Macaulay's Duration}
\usage{
macaulay_duration(face_value = 1000, coupon_rate, int_rate, period)
}
\arguments{
\item{face_value}{Face Value of the Bond, the default value is 1000,}

\item{coupon_rate}{Coupon Payment Rate of the Bond,}

\item{int_rate}{Interest Rate,}

\item{period}{Period of Investment.}
}
\value{
A list with two components

\code{duration} the Macaulay's duration.

\code{calc_data} The required table for calculation.
}
\description{
Given the face value, coupon rate, interest rate (YTM) and the period of investment this function calculates the Macaulay's Duration
}
\examples{
macaulay_duration(coupon_rate=0.16,int_rate=0.17,period=6)
}
\references{
Fabozzi Frank J. and Mann Steven V., Handbook of Fixed Income Securities.
}

