% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pres_val.R
\name{pres_val}
\alias{pres_val}
\title{Present Value Calculation}
\usage{
pres_val(face_value = 1000, int_rate, period)
}
\arguments{
\item{face_value}{Face Value of the security, the default value is 1000,}

\item{int_rate}{Interest Rate,}

\item{period}{Period of Investment.}
}
\value{
The present value of the security.
}
\description{
Given the face value, interest rate and the period of investment this function calculates the present value of a financial security
}
\details{
The formula used for the calculation of the present value is

\deqn{PV=FV\frac{1}{(1+i)^n}}
where FV is the face value and n is the period of investment.
}
\examples{
pres_val(face_value=1000, int_rate=0.17, period=6)
}
\references{
Fabozzi Frank J. and Mann Steven V., Handbook of Fixed Income Securities.
}

