% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vol_b.R
\name{vol_b}
\alias{vol_b}
\title{Volatility of a Bond}
\usage{
vol_b(face_value = 1000, ytm, coupon_rate, period)
}
\arguments{
\item{face_value}{Face Value of the Bond}

\item{ytm}{Yield to Maturity,}

\item{coupon_rate}{Coupon Payment Rate of the Bond,}

\item{period}{Period of Investment.}
}
\value{
The volatility of the bond, subject to the current YTM.
}
\description{
The function calculates the Volatility of a Bond. The Bond in question is assumed to be fixed rate and option-free in nature.
}

